separated TechnicalStatus into its own component, renamed Details to Detail, fixed detail tests
