+++ associate/AssociateDetail.tsx
@@ -4,7 +4,8 @@ import React, { useState } from 'react';
 import 'react-native';
 import { Pressable, View } from 'react-native';
 import { Text, Input, Button, Icon } from 'react-native-elements';
-import associateService, { Associate, QCFeedback } from './associateService';
+import associateService, { Associate, QCFeedback } from './AssociateService';
+import TechnicalStatus from './TechnicalStatus';
 
 interface AssociateProps {
     associate: Associate;
@@ -24,29 +25,12 @@ function AssociateDetail(props: AssociateProps) {
     const [viewNote, setViewNote] = useState(false);
 
     /**
-     * Icon manager(switches the icon based on the value of the technical status);
-     */
-    let iconName: string = 'question';
-    let iconColor: string = 'blue';
-    if (qcTechnicalStatus >= 1 && qcTechnicalStatus < 2) {
-        iconName = 'frown';
-        iconColor = 'red';
-    } else if (qcTechnicalStatus < 3) {
-        iconName = 'meh';
-        iconColor = 'yellow';
-    } else if (qcTechnicalStatus < 4) {
-        iconName = 'smile';
-        iconColor = 'green';
-    }
-
-
-    /**
      * When the Technical Status component is pressed, should cycle through 0-4
      *  Update both the state and the database.
     */
     function cycleTechnicalStatus() {
         let newStatus = qcTechnicalStatus + 1;
-        if (newStatus > 4) {
+        if (newStatus > 5) {
             newStatus = 0;
         }
         setQcTechnicalStatus(newStatus);
@@ -59,10 +43,8 @@ function AssociateDetail(props: AssociateProps) {
             <Text testID='lastName'> {props.associate.lastName}</Text>
             <Pressable
                 onPress={cycleTechnicalStatus}>
-                <Icon
-                    name={iconName}
-                    type='font-awesome'
-                    color={iconColor} />
+                <TechnicalStatus
+                    status={qcTechnicalStatus} />
             </Pressable>
             <Button
                 icon={
@@ -86,8 +68,7 @@ function AssociateDetail(props: AssociateProps) {
                         name='save'
                         type='fontawesome' />
                 }
-            //onPress={() => associateService.updateAssociate({'qcNote': qcNote})}
-            />
+                onPress={() => associateService.updateAssociate(props.qcFeedback, {'qcNote': qcNote})}/>
             }
         </View>
     );
